requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
IPython==8.29.0
openai==1.55.2
python-dotenv==1.0.1
sse_starlette==2.1.3
requests==2.32.3
tiktoken==0.8.0
lxml==5.3.0

pymysql==1.1.1
cryptography==44.0.0
python-multipart==0.0.17